pyproject.toml
[project]
name = "agente-rolplay"
version = "0.1.0"
description = "WhatsApp AI Agent with Google Drive integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiohttp-retry==2.9.1",
    "aiosignal==1.4.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anthropic==0.75.0",
    "anyio==4.11.0",
    "attrs==25.4.0",
    "cachetools==6.2.2",
    "celery>=5.6.2",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "cloudinary>=1.44.1",
    "coloredlogs==15.0.1",
    "distro==1.9.0",
    "docstring-parser==0.17.0",
    "dotenv>=0.9.9",
    "et-xmlfile==2.0.0",
    "fastapi==0.122.0",
    "fastembed==0.7.3",
    "filelock==3.20.0",
    "flatbuffers==25.9.23",
    "frozenlist==1.8.0",
    "fsspec==2025.10.0",
    "google-api-core==2.28.1",
    "google-api-python-client==2.187.0",
    "google-auth==2.41.1",
    "google-auth-httplib2==0.2.1",
    "google-auth-oauthlib==1.2.3",
    "google-cloud-core==2.5.0",
    "google-cloud-firestore==2.21.0",
    "google-cloud-storage==3.6.0",
    "google-crc32c==1.7.1",
    "google-resumable-media==2.8.0",
    "googleapis-common-protos==1.72.0",
    "grpcio==1.76.0",
    "grpcio-status==1.76.0",
    "h11==0.16.0",
    "h2==4.3.0",
    "hf-xet==1.2.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httplib2==0.31.0",
    "httpx==0.28.1",
    "huggingface-hub==0.36.0",
    "humanfriendly==10.0",
    "hyperframe==6.1.0",
    "idna==3.11",
    "jiter==0.12.0",
    "loguru==0.7.3",
    "lxml==6.0.2",
    "mmh3==5.2.0",
    "mpmath==1.3.0",
    "multidict==6.7.0",
    "numpy==2.3.5",
    "oauthlib==3.3.1",
    "onnxruntime==1.23.2",
    "openai==2.8.1",
    "openpyxl==3.1.5",
    "packaging==25.0",
    "pillow==11.3.0",
    "pinecone>=7.0.1",
    "portalocker==3.2.0",
    "propcache==0.4.1",
    "proto-plus==1.26.1",
    "protobuf==6.33.1",
    "py-rust-stemmers==0.1.5",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pydub>=0.25.1",
    "pyjwt==2.10.1",
    "pyparsing==3.2.5",
    "pypdf2==3.0.1",
    "python-docx==1.2.0",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.20",
    "python-pptx==1.0.2",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "qdrant-client==1.16.0",
    "redis==7.1.0",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "rsa==4.9.1",
    "sniffio==1.3.1",
    "starlette==0.50.0",
    "supabase>=2.28.0",
    "sympy==1.14.0",
    "tiktoken==0.12.0",
    "tokenizers==0.22.1",
    "tqdm==4.67.1",
    "twilio==9.8.7",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "uvicorn==0.38.0",
    "xlsxwriter==3.2.9",
    "yarl==1.22.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "jinja2>=3.1.0",
    "itsdangerous>=2.1.0",
    "email-validator>=2.0.0",
]

[project.scripts]
agente-rolplay = "src.agente_rolplay.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src.*"]
